+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/TwitterHashtagButton.js
@@ -1,13 +1,13 @@
 import React, { Fragment } from "react"
 import PropTypes from "prop-types"
 
-export const TwitterHashtagButton = ({ hashtag }) => {
+export const TwitterHashtagButton = ({ hashtag, size }) => {
   return (
     <Fragment>
       <a
         href={`https://twitter.com/intent/tweet?button_hashtag=${hashtag}&ref_src=twsrc%5Etfw`}
         className="twitter-hashtag-button"
-        // data-show-count={true}
+        data-size={size}
       >{`Tweet #${hashtag}`}</a>
     </Fragment>
   )
@@ -18,6 +18,12 @@ TwitterHashtagButton.propTypes = {
    * Twitter hashtag
    */
   hashtag: PropTypes.string.isRequired,
+  /**
+   * The size of the button
+   */
+  size: PropTypes.oneOf(["large", "small"]),
 }
 
-TwitterHashtagButton.defaultProps = {}
+TwitterHashtagButton.defaultProps = {
+  size: "small",
+}