+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/TwitterMentionButton.js
@@ -1,13 +1,13 @@
 import React, { Fragment } from "react"
 import PropTypes from "prop-types"
 
-export const TwitterMentionButton = ({ username }) => {
+export const TwitterMentionButton = ({ username, size }) => {
   return (
     <Fragment>
       <a
         href={`https://twitter.com/intent/tweet?screen_name=${username}&ref_src=twsrc%5Etfw`}
         className="twitter-mention-button"
-        // data-show-count={true}
+        data-size={size}
       >{`Tweet to @${username}`}</a>
     </Fragment>
   )
@@ -18,6 +18,12 @@ TwitterMentionButton.propTypes = {
    * Twitter username
    */
   username: PropTypes.string.isRequired,
+  /**
+   * The size of the button
+   */
+  size: PropTypes.oneOf(["large", "small"]),
 }
 
-TwitterMentionButton.defaultProps = {}
+TwitterMentionButton.defaultProps = {
+  size: "small",
+}