+++ @pauliescanlon/gatsby-mdx-embed/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.9] - 2020-01-xx
+
+## Add
+
+- `<TwitterTimeline />` new component
+- `<TwitterList />` new component
+- `<TwitterFollowButton>` new props `showUsername` and `size`
+- `<TwitterHashtagButton>` new props and `size`
+- `<TwitterMentionButton>` new props and `size`
+
 ## [0.0.8] - 2020-01-14
 
 Technically speaking this patch version does contain breaking changes **but** the first major version hasn't been published yet,